fix(auth): moved auth files to corresponding package
