+++ app/src/main/java/com/android/mySwissDorm/model/authentification/AuthRepositoryFirebase.kt
@@ -1,4 +1,4 @@
-package com.android.mySwissDorm.model
+package com.android.mySwissDorm.model.authentification
 
 import androidx.credentials.Credential
 import androidx.credentials.CustomCredential
